# src/users/users.controller.ts
import {
  Controller,
  Get,
  Post,
  Put,
  Delete,
  Param,
  Body,
  HttpStatus,
  HttpCode,
  ClassSerializerInterceptor,
  UseInterceptors,
} from '@nestjs/common';
import { UsersService } from './users.service';
import { CreateUserDto } from './dto/new-user.dto';
import { UpdateUserDto } from './dto/modify-user.dto';
import { UUIDvalidate } from 'src/UUID.validator';

@UseInterceptors(ClassSerializerInterceptor)
@Controller('user')
export class UsersController {
  constructor(private readonly usersService: UsersService) {}

  @Get()
  @HttpCode(HttpStatus.OK)
  getAllUsers() {
    const users = this.usersService.getAll();
    return users.length ? users : [];
  }

  @Get(':id')
  @HttpCode(HttpStatus.OK)
  getUserById(@Param('id', UUIDvalidate) id: string) {
    return this.usersService.getById(id);
  }

  @Post()
  @HttpCode(HttpStatus.CREATED)
  createUser(@Body() createUserDto: CreateUserDto) {
    return this.usersService.create(createUserDto);
  }

  @Put(':id')
  @HttpCode(HttpStatus.OK)
  updateUser(
    @Param('id', UUIDvalidate) id: string,
    @Body() updatePasswordDto: UpdateUserDto,
  ) {
    return this.usersService.update(id, updatePasswordDto);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  deleteUser(@Param('id', UUIDvalidate) id: string) {
    return this.usersService.delete(id);
  }
}